[#27] create controller for authentication related endpoints
